# Add vectorized Adler32