@voided 2025-09-17
Does the Six Labors Split License have any downstream effects for end users using SteamKit commercially?