@xPaw 2025-09-19
Internals are only exposed to tests in debug.